+++ src/wp-content/mu-plugins/uq-custom-sitewide.php
@@ -21,6 +21,9 @@ define('UQ_BLOG_KIT_PATH', plugin_dir_path(__FILE__));
 
 // 메인 클래스 파일 로드
 require_once UQ_BLOG_KIT_PATH . 'uq-custom-sitewide/includes/uq-blog-kit.php';
+require_once UQ_BLOG_KIT_PATH . 'uq-custom-sitewide/includes/uq-gf-general.php';
+require_once UQ_BLOG_KIT_PATH . 'uq-custom-sitewide/includes/sidebar-cat-filter-widget.php';
+require_once UQ_BLOG_KIT_PATH . 'uq-custom-sitewide/includes/related-posts-section.php';
 
 // 플러그인 초기화
 add_action('init', function() {